Updated to SWIG 1.3.24 (plus a patch that corrects a bug and adds back some things that were changed/removed from my patch I submitted to them.)
Since it is now possible easily and simply share the SWIG type tables
across modules I reverted to always using the stock SWIG runtime
instead of my slightly hacked up version of it exported via the
wxPython C API.

The %name directive is now deprecated so replaced most uses of it with
a custom %Rename macro that uses %rename internally.  These will
evetually need to be replaced with a DocDecl macro when docstrings are
added.


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk/wxPython@31128 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775