Use generic wxAboutBox() in wxUniv/GTK.
The native GTK implementation is not used in wxUniv/GTK, so don't exclude the
generic one as well.

Notice that the same should be done for wxMSW and wxOSX too but we currently
seem to use the native version even in wxUniv there. It would be better to
consistently use the generic version in all wxUniv ports.

Closes #14236.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@71309 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775