Don't define _CRTBLD when using MSVC 12 or later.
This results in fatal #errors from the standard <assert.h> which doesn't want
to be included with _CRTBLD defined. Generally speaking, defining _CRTBLD
looks like a very wrong thing to do and it's probably not necessary any more
(considering that it was done in r25683 almost 10 years ago) but only avoid
doing this with VC 12 for now to avoid breaking anything else before 3.0.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_3_0_BRANCH@75093 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775